requirements.txt
Flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
plotly>=5.18.0
datasets>=2.16.0
transformers>=4.36.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.14.0